+++ feature/main/src/main/java/com/teamoffroad/feature/main/component/MainNavHost.kt
@@ -8,11 +8,11 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
-import com.teamoffroad.feature.auth.navigation.authNavGraph
-import com.teamoffroad.feature.auth.navigation.setBirthDateNavGraph
-import com.teamoffroad.feature.auth.navigation.setCharacterNavGraph
-import com.teamoffroad.feature.auth.navigation.setGenderNavGraph
-import com.teamoffroad.feature.auth.navigation.setNicknameNavGraph
+import com.teamoffroad.feature.auth.presentation.navigation.authNavGraph
+import com.teamoffroad.feature.auth.presentation.navigation.setBirthDateNavGraph
+import com.teamoffroad.feature.auth.presentation.navigation.setCharacterNavGraph
+import com.teamoffroad.feature.auth.presentation.navigation.setGenderNavGraph
+import com.teamoffroad.feature.auth.presentation.navigation.setNicknameNavGraph
 import com.teamoffroad.feature.explore.presentation.navigation.exploreNavGraph
 import com.teamoffroad.feature.home.navigation.homeNavGraph
 import com.teamoffroad.feature.main.MainNavigator
